services/agent-connector/requirements.txt
fastapi
httpx
pydantic
pyyaml
jsonschema    # optional; wenn nicht installierbar, Validierung überspringen
